requirements.txt
Flask
matplotlib
mplsoccer
pandas
Pillow
gunicorn
datetime
request
bs4